+++ mirrord-layer/src/lib.rs
@@ -2,13 +2,10 @@
 #![feature(once_cell)]
 #![feature(result_option_inspect)]
 #![feature(const_trait_impl)]
-#![feature(type_alias_impl_trait)]
-#![feature(generic_associated_types)]
 
 use std::{
-    env,
-    lazy::{SyncLazy, SyncOnceCell},
-    sync::Mutex,
+    collections::HashSet,
+    sync::{LazyLock, Mutex, OnceLock},
 };
 
 use actix_codec::{AsyncRead, AsyncWrite};
@@ -23,9 +20,10 @@ use futures::{SinkExt, StreamExt};
 use kube::api::Portforwarder;
 use libc::c_int;
 use mirrord_protocol::{
-    ClientCodec, ClientMessage, CloseFileRequest, CloseFileResponse, DaemonMessage, FileRequest,
-    FileResponse, OpenFileRequest, OpenFileResponse, OpenRelativeFileRequest, ReadFileRequest,
-    ReadFileResponse, SeekFileRequest, SeekFileResponse, WriteFileRequest, WriteFileResponse,
+    ClientCodec, ClientMessage, CloseFileRequest, CloseFileResponse, DaemonMessage, EnvVars,
+    FileRequest, FileResponse, GetEnvVarsRequest, OpenFileRequest, OpenFileResponse,
+    OpenRelativeFileRequest, ReadFileRequest, ReadFileResponse, SeekFileRequest, SeekFileResponse,
+    WriteFileRequest, WriteFileResponse,
 };
 use sockets::SOCKETS;
 use tcp::TcpHandler;
@@ -52,13 +50,14 @@ mod sockets;
 mod tcp;
 mod tcp_mirror;
 
-use crate::{common::HookMessage, config::Config, macros::hook};
+use crate::{common::HookMessage, config::LayerConfig, macros::hook};
 
-static RUNTIME: SyncLazy<Runtime> = SyncLazy::new(|| Runtime::new().unwrap());
-static GUM: SyncLazy<Gum> = SyncLazy::new(|| unsafe { Gum::obtain() });
+static RUNTIME: LazyLock<Runtime> = LazyLock::new(|| Runtime::new().unwrap());
+static GUM: LazyLock<Gum> = LazyLock::new(|| unsafe { Gum::obtain() });
 
 pub static mut HOOK_SENDER: Option<Sender<HookMessage>> = None;
-pub static ENABLED_FILE_OPS: SyncOnceCell<bool> = SyncOnceCell::new();
+
+pub static ENABLED_FILE_OPS: OnceLock<bool> = OnceLock::new();
 
 #[ctor]
 fn init() {
@@ -69,18 +68,10 @@ fn init() {
 
     info!("Initializing mirrord-layer!");
 
-    let config = Config::init_from_env().unwrap();
+    let config = LayerConfig::init_from_env().unwrap();
 
     let port_forwarder = RUNTIME
-        .block_on(pod_api::create_agent(
-            &config.impersonated_pod_name,
-            &config.impersonated_pod_namespace,
-            &config.agent_namespace,
-            config.agent_rust_log,
-            config.agent_image.unwrap_or_else(|| {
-                concat!("ghcr.io/metalbear-co/mirrord:", env!("CARGO_PKG_VERSION")).to_string()
-            }),
-        ))
+        .block_on(pod_api::create_agent(config.clone()))
         .unwrap();
 
     let (sender, receiver) = channel::<HookMessage>(1000);
@@ -91,7 +82,7 @@ fn init() {
     let enabled_file_ops = ENABLED_FILE_OPS.get_or_init(|| config.enabled_file_ops);
     enable_hooks(*enabled_file_ops);
 
-    RUNTIME.spawn(poll_agent(port_forwarder, receiver));
+    RUNTIME.spawn(poll_agent(port_forwarder, receiver, config));
 }
 
 #[allow(clippy::too_many_arguments)]
@@ -101,7 +92,6 @@ async fn handle_hook_message(
     codec: &mut actix_codec::Framed<impl AsyncRead + AsyncWrite + Unpin + Send, ClientCodec>,
     // TODO: There is probably a better abstraction for this.
     open_file_handler: &Mutex<Vec<oneshot::Sender<OpenFileResponse>>>,
-    open_relative_file_handler: &Mutex<Vec<oneshot::Sender<OpenFileResponse>>>,
     read_file_handler: &Mutex<Vec<oneshot::Sender<ReadFileResponse>>>,
     seek_file_handler: &Mutex<Vec<oneshot::Sender<SeekFileResponse>>>,
     write_file_handler: &Mutex<Vec<oneshot::Sender<WriteFileResponse>>>,
@@ -146,10 +136,7 @@ async fn handle_hook_message(
                 relative_fd, path, open_options
             );
 
-            open_relative_file_handler
-                .lock()
-                .unwrap()
-                .push(file_channel_tx);
+            open_file_handler.lock().unwrap().push(file_channel_tx);
 
             let open_relative_file_request = OpenRelativeFileRequest {
                 relative_fd,
@@ -279,7 +266,7 @@ async fn handle_daemon_message(
         }
         DaemonMessage::FileResponse(FileResponse::Open(open_file)) => {
             debug!("DaemonMessage::OpenFileResponse {open_file:#?}!");
-
+            debug!("file handler = {:#?}", open_file_handler);
             open_file_handler
                 .lock()
                 .unwrap()
@@ -346,12 +333,37 @@ async fn handle_daemon_message(
                 panic!("Daemon: unmatched pong!");
             }
         }
+        DaemonMessage::GetEnvVarsResponse(remote_env_vars) => {
+            debug!("DaemonMessage::GetEnvVarsResponse {:#?}!", remote_env_vars);
+
+            match remote_env_vars {
+                Ok(remote_env_vars) => {
+                    for (key, value) in remote_env_vars.into_iter() {
+                        debug!(
+                            "DaemonMessage::GetEnvVarsResponse set key {:#?} value {:#?}",
+                            key, value
+                        );
+
+                        std::env::set_var(&key, &value);
+                        debug_assert_eq!(std::env::var(key), Ok(value));
+                    }
+                }
+                Err(fail) => error!(
+                    "Loading remote environment variables failed with {:#?}",
+                    fail
+                ),
+            }
+        }
         DaemonMessage::Close => todo!(),
         DaemonMessage::LogMessage(_) => todo!(),
     }
 }
 
-async fn poll_agent(mut pf: Portforwarder, mut receiver: Receiver<HookMessage>) {
+async fn poll_agent(
+    mut pf: Portforwarder,
+    mut receiver: Receiver<HookMessage>,
+    config: LayerConfig,
+) {
     let port = pf.take_stream(61337).unwrap(); // TODO: Make port configurable
 
     // `codec` is used to retrieve messages from the daemon (messages that are sent from -agent to
@@ -367,13 +379,42 @@ async fn poll_agent(mut pf: Portforwarder, mut receiver: Receiver<HookMessage>) 
     // Stores a list of `oneshot`s that communicates with the hook side (send a message from -layer
     // to -agent, and when we receive a message from -agent to -layer).
     let open_file_handler = Mutex::new(Vec::with_capacity(4));
-    let open_relative_file_handler = Mutex::new(Vec::with_capacity(4));
     let read_file_handler = Mutex::new(Vec::with_capacity(4));
     let seek_file_handler = Mutex::new(Vec::with_capacity(4));
     let write_file_handler = Mutex::new(Vec::with_capacity(4));
     let close_file_handler = Mutex::new(Vec::with_capacity(4));
 
     let mut ping = false;
+
+    if !config.override_env_vars_exclude.is_empty() && !config.override_env_vars_include.is_empty()
+    {
+        panic!(
+            r#"mirrord-layer encountered an issue:
+
+            mirrord doesn't support specifying both
+            `OVERRIDE_ENV_VARS_EXCLUDE` and `OVERRIDE_ENV_VARS_INCLUDE` at the same time!
+
+            > Use either `--override_env_vars_exclude` or `--override_env_vars_include`.
+            >> If you want to include all, use `--override_env_vars_include="*"`."#
+        );
+    } else {
+        let env_vars_filter = HashSet::from(EnvVars(config.override_env_vars_exclude));
+        let env_vars_select = HashSet::from(EnvVars(config.override_env_vars_include));
+
+        if !env_vars_filter.is_empty() || !env_vars_select.is_empty() {
+            let codec_result = codec
+                .send(ClientMessage::GetEnvVarsRequest(GetEnvVarsRequest {
+                    env_vars_filter,
+                    env_vars_select,
+                }))
+                .await;
+
+            debug!(
+                "ClientMessage::GetEnvVarsRequest codec_result {:#?}",
+                codec_result
+            );
+        }
+    }
 
     let mut tcp_mirror_handler = TcpMirrorHandler::default();
 
@@ -384,7 +425,6 @@ async fn poll_agent(mut pf: Portforwarder, mut receiver: Receiver<HookMessage>) 
                 &mut tcp_mirror_handler,
                 &mut codec,
                 &open_file_handler,
-                &open_relative_file_handler,
                 &read_file_handler,
                 &seek_file_handler,
                 &write_file_handler,