+++ CHANGELOG.md
@@ -7,7 +7,12 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+
+- Added support for overriding a process' environment variables by setting `MIRRORD_OVERRIDE_ENV_VARS` to `true`. To filter out undesired variables, use the `MIRRORD_OVERRIDE_FILTER_ENV_VARS` configuration with arguments such as `FOO;BAR`.
+
 ### Changed
+
 - Removed `unwrap` from the `Future` that was waiting for Kube pod to spin up in `pod_api.rs`. (Fixes #110)
 - Speed up agent container image building by using a more specific base image.
 - CI: Remove building agent before building & running tests (duplicate)
@@ -16,19 +21,30 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Refactor: Change protocol + mirrord-layer to split messages into modules, so main module only handles general messages, passing down to the appropriate module for handling.
 - Refactor(agent) - change `FileManager` to be per peer, thus removing the need of it being in a different task, moving the handling to the peer logic, change structure of peer handling to a struct.
 - Add a CLI flag to specify `MIRRORD_AGENT_TTL`
+- CI: Collect mirrord-agent logs in case of failure in e2e.
+- Add "app" = "mirrord" label to the agent pod for log collection at ease.
+- CI: Add sleep after local app finishes loading for agent to load filter make tests less flaky.
+- Handle relative paths for open, openat
+- Fix once cell renamings, PR [#98165](https://github.com/rust-lang/rust/pull/98165)
+- Enable the blocking feature of the `reqwest` library 
 
 ## 2.2.1
+
 ### Changed
+
 - Compile universal binaries for MacOS. (Fixes #131)
 - E2E small improvements, removing sleeps. (Fixes #99)
 
 ## 2.2.0
+
 ### Added
+
 - File operations are now available behind the `MIRRORD_FILE_OPS` env variable, this means that mirrord now hooks into the following file functions: `open`, `fopen`, `fdopen`, `openat`, `read`, `fread`, `fileno`, `lseek`, and `write` to provide a mirrored file system.
-- Support for running x64 (Intel) binary on arm (Silicon) macOS using mirrord. This will download and use the x64 mirrord-layer binary when needed.   
+- Support for running x64 (Intel) binary on arm (Silicon) macOS using mirrord. This will download and use the x64 mirrord-layer binary when needed.
 - Add detours for fcntl/dup system calls, closes [#51](https://github.com/metalbear-co/mirrord/issues/51)
 
 ### Changed
+
 - Add graceful exit for library extraction logic in case of error.
 - Refactor the CI by splitting the building of mirrord-agent in a separate job and caching the agent image for E2E tests.
 - Update bug report template to apply to the latest version of mirrord.
@@ -42,6 +58,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Fix typos
 
 ## 2.1.0
+
 ### Added
 
 - Prompt user to update if their version is outdated in the VS Code extension or CLI.
@@ -49,6 +66,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add a keep-alive to keep the agent-pod from exiting, closes [#63](https://github.com/metalbear-co/mirrord/issues/63)
 
 ## 2.0.4
+
 Complete refactor and re-write of everything.
 
 - The CLI/VSCode extension now use `mirrord-layer` which loads into debugged process using `LD_PRELOAD`/`DYLD_INSERT_LIBRARIES`.